Add splitFqdn and splitFqdnInZone
These functions extract a record name and zone (represented by a
RnameZone) object from a fully-qualified domain name. The former does
not require a zone to be specified, and determines it based on the
publicsuffix.org list of top-level domains. The latter *does* require a
zone name to be specified, and is intended for use in situations where
you already know the zone, but really just want to get the record name.

This commit also makes the RnameZone class iterable, so you can do:

    (rname,zone) = splitFqdn('www.bbc.co.uk')

If you only want to access a single component, you can use the
appropriate attribute of the result:

    print(splitFqdn('www.bbc.co.uk').rname) # 'www'
    print(splitFqdn('www.bbc.co.uk').zone)  # 'bbc.co.uk.'