Store resource tree in JSON representation
For all ResourceData objects that have their rtree attribute set,
indicating the presence of load balancing information, store this in the
JSON serialisation of the zone. Similarly, decode this info when
creating a ZoneData object from a dict.

Where resources exists which are CNAMEs used for load balancing
purposes, these are excluded from the main list of resources, since
their information is present in the resource tree and those resources
can be reconstructed from there.