Allow multiple CNAMEs for a given resource
CNAMEs are actually supposed to be singletons, at least in the context
of the protocol. However when load balancing is configured for a
resource, there can be multiple CNAME records for a given resource, each
representing one possible choice. Only of of these is ever returned to
resolvers at query time, but we need to represent multiple CNAMEs when
interacting with vendors.