Add ResolutionPath
This class represents a path through a resource/resolution tree from the
root to a leaf, e.g. [geo region, weighted set, record].

RecordSets are excluded from the paths as they are not relevant; the
need or lack thereof for a RecordSet depends only on whether or not
there are multiple records at the relevant place in the tree.

ResourceTree now has a paths property, which returns a set of all paths.
These sets can be used to determine a delta between two trees, by
checking which paths are present in one but absent from another.