Fix setup.py to work with pip 6.0.8
setup.py was using the parse_requirements function from pip to populate
the install_requires data based on the contents of requirements.txt.

It turns out we're not actually supposed to do this, as the pip API
isn't public, and attempting to set up the package when using the
current version of pip (6.0.8) doesn't work as they've changed their
(private) API so that the way it was being used here doesn't work any
more. See:

https://mail.python.org/pipermail/distutils-sig/2015-February/025782.html

So now we just use [] for the install_requires (that is, no dependencies
listed), and instead rely on the assumption that

    pip install -r requirements.txt

has already been run.

See here for a discussion of the difference between the two ways of
specifying dependencies:

https://caremad.io/2013/07/setup-vs-requirement/